backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
bibtexparser
pydantic[email]
pyjwt
fastapi-mail
aiosmtplib
python-multipart
pytest
pytest-cov
pytest-asyncio
httpx